+++ NEWS.md
@@ -1,5 +1,9 @@
 # openxlsx2 (development version)
 
+## Fixes
+
+* `wb_dims()` no longer slows down when called with large data frames. [#1624](https://github.com/JanMarvin/openxlsx2/pull/1624)
+
 
 ***************************************************************************